infra/requirements.txt
aws-cdk.core==1.113.0
aws_cdk.aws_ecs==1.113.0
aws_cdk.aws_rds==1.113.0
aws_cdk.aws_secretsmanager==1.113.0
aws-cdk.aws-ecr==1.113.0
aws-cdk.aws-iam==1.113.0
aws-cdk.aws-apprunner==1.113.0
aws-cdk.aws-autoscaling==1.113.0